Explicitly set GMT when doing entitlement date math
There were two places that were using non-GMT dates for entitlement math. The
first was in test_validity when making cert dates. The second was in
validity.py, in _get_entitlements_spanning_now.

The time zone attribute on the date was being altered, but without affecting
the date itself. For example, see https://gist.github.com/3666602.