Add plugin for checking if repos were modified during downloads.
When we use hourly certs, it is possible for a certificate to expire in the
middle of a download. If this happens, the current downloading package will
complete, but subsequent downloads will fail.

The fix for this is to add a new plugin that stores the modification times of
the repo files before downloads begin, and then if there were download issues
*and* the repo files were altered, tell the user to re-run their yum command.
This will let the user pick up where they left off with the download.

To test, I used the "trickle" package from EPEL to cap yum to 10KB/s, and then
downloaded a large package set.