If product cert is found, ensure it is in product db.
If a product cert is on the system, it may or not be in the product db. For
example, this occurs on systems that had their product cert copied into place
after installation of packages.

When checking if a product cert exists, also check that it is in the product
db as well, and fix the product db if needed.