New plugin to erase all packages associated with a product.
There are cases when a user will want to remove a product from a system. Two
examples are if the user wants to get "green" without adding entitlements, or
if the user is consuming more products than they'd like via splice.

This plugin allows the user to run "yum erase-product --productname 'foo'", and
all packages installed from the 'foo' repo will be removed. This will cause the
product-id plugin to recognize that 'foo' is no longer being consumed, and the
product cert will be removed from the system.